fix: tabs layout and tabs component
