+++ messages/sv.json
@@ -232,14 +232,14 @@
     "logIn": "Logga in",
     "logOut": "Logga ut",
     "signIn": "Logga in",
-    "signUp": "Registrera dig",
+    "signUp": "Skapa konto",
     "signOut": "Logga ut",
     "loading": "Laddar...",
     "button": "Skicka",
     "loggedIn": "Du är inloggad!",
-    "enterEmail": "Skriv din email",
-    "enterPassword": "Skriv ditt lösenord",
-    "confirmPassword": "Bekräfta lösenordet",
+    "enterEmail": "Ange e-postadress",
+    "enterPassword": "Ange lösenord",
+    "confirmPassword": "Bekräfta lösenord",
     "forgotPassword": "Glömt ditt lösenord?",
     "firstName": "Förnamn",
     "lastName": "Efternamn",
