+++ src/app/[locale]/(tabs)/user/profile/page.tsx
