--- src/app/[locale]/user/reset-password/page.tsx
@@ -1,69 +1,0 @@
-'use client';
-import {useState} from 'react';
-import {Button} from '@/components/ui/button';
-import {Input} from '@/components/ui/input';
-import {supabase} from '@/lib/supabase/supabaseClient';
-import {useTranslations} from 'next-intl';
-import { useToast } from "@/hooks/useToast"
-
-const ResetPassword = () => {
-  const { toast } = useToast();
-  const t = useTranslations('authenticate');
-  const [loading, setLoading] = useState(false);
-  const [userEmail, setUserEmail] = useState('');
-  const handleSubmit = async (e: React.FormEvent) => {
-    e.preventDefault();
-    setLoading(true);
-    try {
-      const { error } = await supabase.auth.resetPasswordForEmail(userEmail);
-      if (error) throw error;
-      toast({
-        title: t('emailSent'),
-        description: t('emailSentDescription'),
-      });
-      setUserEmail('');
-    } catch (error) {
-      console.error('Error resetting password:', error);
-      toast({
-        title: t('error'),
-        description: t('errorDescription'),
-        variant: 'destructive',
-      });
-    } finally {
-      setLoading(false);
-    }
-  };
-
-  const handleChange = (e: React.ChangeEvent<HTMLInputElement>) => {
-    const {value} = e.target;
-    setUserEmail(value);
-  };
-
-  return (
-    <form onSubmit={handleSubmit} className="flex h-fit min-h-[448px] flex-col justify-between max-w-[600px] mx-auto mt-10">
-      <div className="space-y-6 px-4">
-        <h2 className="text-2xl font-bold text-center">{t('resetPassword')}</h2>
-        <div className="space-y-2"> 
-          <Input
-            id="login-email"
-            name="email"
-            type="email"
-            placeholder={t('enterEmail')}
-            value={userEmail}
-            onChange={handleChange}
-            required
-            className="h-12 rounded-full"
-            disabled={loading}
-          />
-        </div>
-      </div>
-      <div className="mt-14 flex justify-center pb-6">
-        <Button type="submit" disabled={loading} variant="pink">
-          {loading ? t('loading') : t('reset')}
-        </Button>
-      </div>
-    </form>
-  );
-};
-
-export default ResetPassword;
