+++ src/components/index.ts
@@ -1,12 +1,13 @@
-export * from './ui';
-export * from './LanguageSelector';
-export * from './Timer';
-export * from './RandomQuestion';
-export * from './PassItOn';
-export * from './HomeButton';
-export * from './RiveAnimation';
 export * from './About';
-export * from './ExerciseCard';
-export * from './PageLayout';
+export * from './AppBar';
 export * from './CarouselPagination';
+export * from './ExerciseCard';
 export * from './Header';
+export * from './HomeButton';
+export * from './LanguageSelector';
+export * from './PageLayout';
+export * from './PassItOn';
+export * from './RandomQuestion';
+export * from './RiveAnimation';
+export * from './Timer';
+export * from './ui';
